requirements.txt
azureml-sdk
click